Remove auto-update, added queued items for delayed add (experimental)
Removed auto-update
Added priority 1, to load before (most) other mods
if RadialMouseMenu:new() is called before game load but after RMM loads, creation of the menu will be delayed until the game loads.
If a second argument `callback` is supplied to new(), it will be called upon successful radial menu instance creation with the newly created menu as an argument.